Use a common BERT request builder
